fix(pwa): avoid duplicate party settings history
Summary:
- Share the Back button navigation behavior through a reusable hook.
- Return from saved settings subsections through browser history.
- Cover one-click navigation back to the party expense list.

Rationale:
- Replacing a subsection with the settings hub left duplicate hub entries,
  forcing users to press Back more than once.

Tests:
- vp run check
- vp run test
- mise exec -- vp run build
- vp exec playwright test e2e/settings.spec.ts
- focused history tests repeated five times in Chromium
